app/requirements.txt
streamlit
requests
openai
phidata
groq
duckduckgo_search
pypdf
arxiv 